pyproject.toml
[tool.poetry]
name = "xcellaccession"
version = "0.1.0"
description = ""
authors = ["srbayless <mrssrwright@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.2.9"
django-phone-field = "^1.8.1"
pillow = "^8.4.0"
django-rest-framework = "^0.1.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
